+++ src/machine.rs
@@ -2,6 +2,7 @@
 
 use std::{
     collections::{BTreeMap, HashSet},
+    io::{Cursor, Read},
     iter,
     ops::Deref,
     pin::{pin, Pin},
@@ -11,23 +12,30 @@ use std::{
 use futures_util::{pin_mut, Stream, StreamExt};
 use js_sys::{Array, Function, JsString, Map, Promise, Set};
 use matrix_sdk_common::ruma::{
-    self, events::secret::request::SecretName, serde::Raw, OneTimeKeyAlgorithm, OwnedDeviceId,
-    OwnedTransactionId, OwnedUserId, UInt,
+    self,
+    events::{
+        room::{EncryptedFile, EncryptedFileInit},
+        secret::request::SecretName,
+    },
+    serde::Raw,
+    OneTimeKeyAlgorithm, OwnedDeviceId, OwnedTransactionId, OwnedUserId, UInt,
 };
 use matrix_sdk_crypto::{
     backups::MegolmV1BackupKey,
     olm::{BackedUpRoomKey, ExportedRoomKey},
     store::types::{DeviceChanges, IdentityChanges},
-    types::RoomKeyBackupInfo,
-    CryptoStoreError, EncryptionSyncChanges, GossippedSecret,
+    types::{events::room_key_bundle::RoomKeyBundleContent, RoomKeyBackupInfo},
+    CryptoStoreError, EncryptionSyncChanges, GossippedSecret, MediaEncryptionInfo,
 };
 use serde::{ser::SerializeSeq, Serialize, Serializer};
 use serde_json::json;
-use tracing::{dispatcher, instrument::WithSubscriber, warn, Dispatch};
+use tracing::{dispatcher, info, instrument::WithSubscriber, warn, Dispatch};
 use wasm_bindgen::{convert::TryFromJsValue, prelude::*};
 use wasm_bindgen_futures::{spawn_local, JsFuture};
+use zeroize::Zeroizing;
 
 use crate::{
+    attachment,
     backup::{BackupDecryptionKey, BackupKeys, RoomKeyCounts},
     dehydrated_devices::DehydratedDevices,
     device, encryption,
@@ -42,7 +50,7 @@ use crate::{
     tracing::{logger_to_dispatcher, JsLogger},
     types::{
         self, processed_to_device_event_to_js_value, RoomKeyImportResult, RoomSettings,
-        SignatureVerification,
+        SignatureVerification, StoredRoomKeyBundleData,
     },
     verification, vodozemac,
 };
@@ -1679,6 +1687,184 @@ impl OlmMachine {
     pub fn dehydrated_devices(&self) -> DehydratedDevices {
         let _guard = dispatcher::set_default(&self.tracing_subscriber);
         self.inner.dehydrated_devices().into()
+    }
+
+    /// Assemble, and encrypt, a room key bundle for sharing encrypted history,
+    /// as per {@link https://github.com/matrix-org/matrix-spec-proposals/pull/4268|MSC4268}.
+    ///
+    /// Returns `undefined` if there are no keys to share in the given room,
+    /// otherwise an {@link EncryptedAttachment}.
+    ///
+    /// The data should be uploaded to the media server, and the details then
+    /// passed to {@link shareRoomKeyBundleData}.
+    ///
+    /// @experimental
+    #[wasm_bindgen(
+        js_name = "buildRoomKeyBundle",
+        unchecked_return_type = "Promise<EncryptedAttachment | undefined>"
+    )]
+    pub fn build_room_key_bundle(&self, room_id: &identifiers::RoomId) -> Promise {
+        let _guard = dispatcher::set_default(&self.tracing_subscriber);
+        let me = self.inner.clone();
+        let room_id = room_id.inner.clone();
+
+        future_to_promise(async move {
+            let bundle = me.store().build_room_key_bundle(&room_id).await?;
+
+            if bundle.is_empty() {
+                info!("No keys to share");
+                return Ok(None);
+            }
+
+            // Remember to zeroize the json as soon as we're done with it
+            let json = Zeroizing::new(serde_json::to_vec(&bundle)?);
+
+            Ok(Some(attachment::Attachment::encrypt(json.as_slice())?))
+        })
+    }
+
+    /// Collect the devices belonging to the given user, and send the details
+    /// of a room key bundle to those devices.
+    ///
+    /// Returns a list of to-device requests which must be sent.
+    ///
+    /// @experimental
+    #[wasm_bindgen(
+        js_name = "shareRoomKeyBundleData",
+        unchecked_return_type = "Promise<ToDeviceRequest[]>"
+    )]
+    pub fn share_room_key_bundle_data(
+        &self,
+        user: &identifiers::UserId,
+        room: &identifiers::RoomId,
+        url: &str,
+        media_encryption_info: Option<String>,
+        sharing_strategy: encryption::CollectStrategy,
+    ) -> Result<Promise, JsError> {
+        let _guard = dispatcher::set_default(&self.tracing_subscriber);
+        let me = self.inner.clone();
+        let user_id = user.inner.clone();
+
+        let media_encryption_info = media_encryption_info.ok_or_else(|| {
+            // We accept Option<String> to save a typescript assertion on the application
+            // side. In practice `build_room_key_bundle` should never return an
+            // EncryptedAttachment with nullish encryption_info.
+            JsError::new("shareRoomKeyBundleData: nullish encryption info")
+        })?;
+
+        let media_encryption_info: MediaEncryptionInfo =
+            serde_json::from_str(&media_encryption_info).map_err(|e| {
+                JsError::new(&format!("Unable to validate room key media encryption info: {e}"))
+            })?;
+
+        let url = url.try_into().map_err(|e| JsError::new(&format!("Invalid media url: {e}")))?;
+
+        let bundle_data = RoomKeyBundleContent {
+            room_id: room.inner.clone(),
+            file: EncryptedFile::from(EncryptedFileInit {
+                url,
+                key: media_encryption_info.key,
+                iv: media_encryption_info.iv,
+                hashes: media_encryption_info.hashes,
+                v: media_encryption_info.version,
+            }),
+        };
+
+        Ok(future_to_promise(async move {
+            let to_device_requests = me
+                .share_room_key_bundle_data(&user_id, &sharing_strategy.into(), bundle_data)
+                .await?;
+
+            // convert each request to our own ToDeviceRequest struct, and then wrap it in a
+            // JsValue.
+            //
+            // Then collect the results into a javascript Array, throwing any errors into
+            // the promise.
+            let result = to_device_requests
+                .into_iter()
+                .map(|td| ToDeviceRequest::try_from(&td).map(JsValue::from))
+                .collect::<Result<Array, _>>()?;
+
+            Ok(result)
+        }))
+    }
+
+    /// See if we have received an {@link https://github.com/matrix-org/matrix-spec-proposals/pull/4268|MSC4268}
+    /// room key bundle for the given room from the given user.
+    ///
+    /// Returns either `undefined` if no suitable bundle has been received,
+    /// or an {@link StoredRoomKeyBundleData}, in which case, the bundle
+    /// should be downloaded, and then passed to {@link
+    /// receiveRoomKeyBundle}.
+    ///
+    /// @experimental
+    #[wasm_bindgen(
+        js_name = "getReceivedRoomKeyBundleData",
+        unchecked_return_type = "Promise<StoredRoomKeyBundleData | undefined>"
+    )]
+    pub fn get_received_room_key_bundle_data(
+        &self,
+        room_id: &identifiers::RoomId,
+        inviter: &identifiers::UserId,
+    ) -> Promise {
+        let _guard = dispatcher::set_default(&self.tracing_subscriber);
+        let me = self.inner.clone();
+        let room_id = room_id.inner.clone();
+        let inviter = inviter.inner.clone();
+
+        future_to_promise(async move {
+            let result = me
+                .store()
+                .get_received_room_key_bundle_data(&room_id, &inviter)
+                .await?
+                .map(types::StoredRoomKeyBundleData::from);
+            Ok(result)
+        })
+    }
+
+    /// Import the message keys from a downloaded room key bundle.
+    ///
+    /// After {@link getReceivedRoomKeyBundleData} returns a truthy result, the
+    /// media file should be downloaded and then passed into this method to
+    /// actually do the import.
+    ///
+    /// @experimental
+    #[wasm_bindgen(js_name = "receiveRoomKeyBundle", unchecked_return_type = "Promise<undefined>")]
+    pub fn receive_room_key_bundle(
+        &self,
+        bundle_data: &StoredRoomKeyBundleData,
+        encrypted_bundle: Vec<u8>,
+    ) -> Result<Promise, JsError> {
+        let _guard = dispatcher::set_default(&self.tracing_subscriber);
+
+        let deserialized_bundle = {
+            let mut cursor = Cursor::new(encrypted_bundle.as_slice());
+            let mut decryptor = matrix_sdk_crypto::AttachmentDecryptor::new(
+                &mut cursor,
+                serde_json::from_str(&bundle_data.encryption_info)?,
+            )?;
+
+            let mut decrypted_bundle = Zeroizing::new(Vec::new());
+            decryptor.read_to_end(&mut decrypted_bundle)?;
+
+            serde_json::from_slice(&decrypted_bundle)?
+        };
+
+        let me = self.inner.clone();
+        let bundle_data = bundle_data.clone();
+        Ok(future_to_promise(async move {
+            me.store()
+                .receive_room_key_bundle(
+                    &bundle_data.room_id.inner,
+                    &bundle_data.sender_user.inner,
+                    &bundle_data.sender_data,
+                    deserialized_bundle,
+                    /* TODO: Use the progress listener and expose an argument for it. */
+                    |_, _| {},
+                )
+                .await?;
+            Ok(JsValue::UNDEFINED)
+        }))
     }
 
     /// Shut down the `OlmMachine`.